refactor: limit exports, update imports, consolidate utils, remove refactor'd code
